moved velocity to an independent module
